requirements.txt
azure-storage-blob==12.25.1
azure-storage-queue==12.12.0
python-dotenv==1.1.0
torch==2.7.1
torchvision==0.22.1
torchaudio==2.5.1
numpy==2.2.6
scikit-learn==1.6.1
matplotlib==3.10.3
Pillow==11.2.1
tqdm==4.67.1
scipy==1.15.3 